correct use of ptrdiff_t with nrrdPad
git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@2740 3d70eeeb-363e-0410-a505-8a46323a89f2

+++ src/limn/splineMethods.c
@@ -243,7 +243,7 @@ limnSplineNix(limnSpline *spline) {
 int
 limnSplineNrrdCleverFix(Nrrd *nout, Nrrd *nin, int info, int type) {
   char me[]="limnSplineNrrdCleverFix", err[AIR_STRLEN_MED];
-  int min[3], max[3];
+  ptrdiff_t min[3], max[3];
   size_t N;
   unsigned int wantSize;
   Nrrd *ntmpA, *ntmpB;
